+++ crates/gpui/src/platform/linux/x11/event.rs
@@ -39,6 +39,7 @@ pub(crate) fn modifiers_from_state(state: xproto::KeyButMask) -> Modifiers {
         shift: state.contains(xproto::KeyButMask::SHIFT),
         platform: state.contains(xproto::KeyButMask::MOD4),
         function: false,
+        altgr: state.contains(xproto::KeyButMask::MOD5),
     }
 }
 
@@ -50,6 +51,7 @@ pub(crate) fn modifiers_from_xinput_info(modifier_info: xinput::ModifierInfo) ->
         shift: modifier_info.effective as u16 & ModMask::SHIFT.bits() == ModMask::SHIFT.bits(),
         platform: modifier_info.effective as u16 & ModMask::M4.bits() == ModMask::M4.bits(),
         function: false,
+        altgr: modifier_info.effective as u16 & ModMask::M5.bits() == ModMask::M5.bits(),
     }
 }