+++ crates/gpui/src/platform/mac/window.rs
@@ -6,7 +6,9 @@ use crate::{
     PlatformInput, PlatformWindow, Point, PromptButton, PromptLevel, RequestFrameOptions,
     SharedString, Size, SystemWindowTab, Timer, WindowAppearance, WindowBackgroundAppearance,
     WindowBounds, WindowControlArea, WindowKind, WindowParams, dispatch_get_main_queue,
-    dispatch_sys::dispatch_async_f, platform::PlatformInputHandler, point, px, size,
+    dispatch_sys::dispatch_async_f,
+    platform::{PlatformInputHandler, mac::events::is_altgr},
+    point, px, size,
 };
 use block::ConcreteBlock;
 use cocoa::{
@@ -1088,6 +1090,7 @@ impl PlatformWindow for MacWindow {
 
             let control = modifiers.contains(NSEventModifierFlags::NSControlKeyMask);
             let alt = modifiers.contains(NSEventModifierFlags::NSAlternateKeyMask);
+            let altgr = is_altgr(modifiers);
             let shift = modifiers.contains(NSEventModifierFlags::NSShiftKeyMask);
             let command = modifiers.contains(NSEventModifierFlags::NSCommandKeyMask);
             let function = modifiers.contains(NSEventModifierFlags::NSFunctionKeyMask);
@@ -1098,6 +1101,7 @@ impl PlatformWindow for MacWindow {
                 shift,
                 platform: command,
                 function,
+                altgr,
             }
         }
     }
@@ -2340,7 +2344,7 @@ extern "C" fn do_command_by_selector(this: &Object, _: Sel, _: Sel) {
     drop(lock);
 
     if let Some((keystroke, mut callback)) = keystroke.zip(event_callback.as_mut()) {
-        let prefer_character_input = keystroke.modifiers.alt && !keystroke.is_ascii();
+        let prefer_character_input = keystroke.prefer_character_input();
         let handled = (callback)(PlatformInput::KeyDown(KeyDownEvent {
             keystroke,
             is_held: false,