+++ crates/gpui/src/platform/windows/events.rs
@@ -1522,6 +1522,7 @@ pub(crate) fn current_modifiers() -> Modifiers {
         shift: is_virtual_key_pressed(VK_SHIFT),
         platform: is_virtual_key_pressed(VK_LWIN) || is_virtual_key_pressed(VK_RWIN),
         function: false,
+        altgr: is_virtual_key_pressed(VK_RMENU),
     }
 }