+++ crates/terminal/src/mappings/keys.rs
@@ -316,6 +316,7 @@ mod test {
                 shift: false,
                 platform: false,
                 function: false,
+                altgr: false,
             },
             key: "🖖🏻".to_string(), //2 char string
             key_char: None,