+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Skip link to openssh.org when finding pull request link in git push output (#287)
+- Improve AltGr / Right option detection on Mac OS (#206)
 
 [sumi-light]: https://github.com/LogicSatinn/sumi-light-zed